Support creating new bundles on the CLI
New behavior is that ./bundle-editor foo.yaml
will edit foo.yaml if it's there or create a new bundle for you if it's
not there. If you edit it, it backs up the old one as foo.yaml~.

The -o flag still works to save somewhere else, so:

./bundle-editor realfile.yaml -o realfile.new.yaml

leaves realfile.yaml untouched.

and this works too:
./bundle-editor nonexistentfile.yaml -o foo.yaml

that'll write to foo.yaml and not create anything.

This also incidentally includes a performance improvement for adding
large bundles - it sends one metadata api call for the whole bundle
instead of one per charm.