Duplicate charm info using key that has series
In some cases (like during a juju api deploy call in conjure-up) we
require that a charm ID includes a series, using the bundle's default
series, and then use that to request charm info. This change ensures
that we can find the info for that charm, even if the info request
didn't include the default series.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>